Lots of changes
* Dropped LED's and LASER control
* Renamed DECLINATION to AZIMUTH
* Moved AZIMUTH control hookup into the globe (it needs to spin freely)

I couldn't make the LED's fit and shine without casting shadows inside
the globe. Plus, there wasn't much room, and it required more wires to
make work. No a critical problem, but the challenges outweighed the
benefits. I'm going to water color paint the globe now.

I also removed the LASER control, the manufacturer said don't duty cycle
it, the built in controls take care of it. This isn't supposed to be on
constantly anyway.